minor changes, good checkpoint
tracker works without reid
