+++ server.py
@@ -2,7 +2,7 @@ import socket
 import cv2
 import numpy as np
 import json
-from tracker import create_tracker
+from tracker import create_tracker, display_fps
 import tensorflow.keras.backend as K
 
 class Server():
@@ -24,19 +24,24 @@ class Server():
 
     def communicate(self):
         tracker = None
+        center = None
+        previous_time = 0
+        
         while True:
             img = self._recieve_img()
+            previous_time = display_fps(img, previous_time)
             
             if tracker is not None:
                 img = tracker.track(img, reid_on=False)
-            
+                center = tracker.tracked_to.centroid if tracker.tracked_to is not None else None
+                
             if cv2.waitKey(1) & 0xFF == ord('s'):
                 tracker = create_tracker(img)
                 
-            json_data = {"string": "hello"}
+            json_data = {"center": center}
             self._send_json(json_data)
             
-            cv2.imshow("***TRACKING***", img)
+            cv2.imshow("*** TRACKING ***", img)
             cv2.waitKey(1)
             
     def _send_json(self, json_data):
