+++ test.py
@@ -1,0 +1,4 @@
+import torch
+
+device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+print("Using device:", device)
